requirements.txt
fastapi==0.115.12
faster_whisper==1.1.1
gTTS==2.5.4
langdetect==1.0.9
numpy==2.2.4
protobuf==6.30.2
pydantic==2.11.3
pydub==0.25.1
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
sounddevice==0.5.1
streamlit==1.44.1
torch==2.5.1+cu118
torch==2.6.0+cu118
transformers==4.51.0
